requirements.txt
# Step 1 — Document ingestion & parsing
pymupdf>=1.24.0          # Digital PDF parsing (import as fitz)
pdfplumber>=0.11.0       # Fallback PDF parser
pdf2image>=1.17.0        # Render scanned PDF pages to images
pytesseract>=0.3.13      # OCR wrapper for Tesseract
Pillow>=10.0.0           # Image handling
fastapi>=0.110.0         # API framework
uvicorn[standard]>=0.29  # ASGI server
python-multipart>=0.0.9  # File upload support for FastAPI

# Step 2 (coming next)
# langchain>=0.2.0
# langchain-community>=0.2.0
# sentence-transformers>=3.0.0
# chromadb>=0.5.0

# Dev / testing
pytest>=8.0.0
httpx>=0.27.0            # For testing FastAPI endpoints
